feat(ui5-viewsettings-dialog): migrate to Typescript
